Migrated advent of code cog to new extension format
